feat(lib): move hybrids/prototype to lib
affects:
  @apollo-elements/lib,
  @apollo-elements/hybrids
