+++ packages/hybrids/factories/client.ts
@@ -2,7 +2,8 @@ import type { ApolloClient, NormalizedCacheObject } from '@apollo/client/core';
 import { Descriptor } from 'hybrids';
 
 import { ApolloElementMixin } from '@apollo-elements/mixins/apollo-element-mixin';
-import { apply, getDescriptor } from '../helpers/prototypes';
+import { hookElementIntoHybridsCache } from '../helpers/cache';
+import { apply, getDescriptor } from '@apollo-elements/lib/prototypes';
 
 export class ApolloElementElement<D = unknown, V = unknown>
   extends ApolloElementMixin(HTMLElement)<D, V> { }
@@ -17,7 +18,7 @@ export function client<TData, TVariables>(
 ): Descriptor<ApolloElementElement<TData, TVariables>> {
   return {
     connect(host) {
-      apply(host, ApolloElementElement, 'client');
+      apply(host, ApolloElementElement, 'client', hookElementIntoHybridsCache);
       host.client = (client ?? (useGlobal && window.__APOLLO_CLIENT__)) || null;
       getDescriptor(host).connectedCallback.value.call(host);
     },
