+++ packages/hybrids/factories/mutation.ts
@@ -2,7 +2,8 @@ import type { DocumentNode } from 'graphql';
 import type { Descriptor } from 'hybrids';
 
 import { ApolloMutationMixin } from '@apollo-elements/mixins/apollo-mutation-mixin';
-import { apply, getDescriptor, hookIntoHybridsRender } from '../helpers/prototypes';
+import { hookPropertyIntoHybridsCache } from '../helpers/cache';
+import { apply, getDescriptor } from '@apollo-elements/lib/prototypes';
 
 class ApolloMutationElement<D = unknown, V = unknown>
   extends ApolloMutationMixin(HTMLElement)<D, V> { }
@@ -26,7 +27,7 @@ export function mutation<TData, TVariables>(
 
     connect(host, key, invalidate) {
       apply(host, ApolloMutationElement, 'mutation');
-      hookIntoHybridsRender({ host, key: 'called', init: false });
+      hookPropertyIntoHybridsCache({ host, key: 'called', init: false });
       // @ts-expect-error: gotta hook up spies somehow
       host?.__testingEscapeHatch?.(host);
       const mo = new MutationObserver(() => invalidate());
